change getUserReviewsAction in spots.js to getUserSpotsAction and change all file import/exports accordingly
