+++ frontend/src/components/Spots/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { NavLink } from "react-router-dom";
 import SpotItem from "./SpotItem";
