+++ frontend/src/store/reviews.js
@@ -1,5 +1,10 @@
 import { csrfFetch } from "./csrf";
 
+// reviews/current GET
+// reviews/${review.id}/images POST
+// reviews/${review.id} DELETE
+// review-images/{reviewImage.id} DELETE
+
 const initialState = { spot: {}, user: {} };
 const reviewsReducer = (state = initialState, action) => {
   let newState;
